chore: adding examples directory and svgs
