+++ README.md
@@ -185,6 +185,9 @@ setInterval(() => createAsciiCharts(), 500);
 
 ## percentage example
 using percentage, solid color with label colors   
+
+<img width="1000" src="https://tool3.github.io/chartscii/img/percentage.svg">   
+
 ```js
 const Chartscii = require('chartscii');
