+++ docs/README.md
@@ -134,6 +134,9 @@ setInterval(() => createAsciiCharts(), 500);
 
 ## fancy example
 fancy example, using labels with colors   
+
+<img width="1000" src="https://tool3.github.io/chartscii/img/fancy.svg">   
+
 ```js
 const Chartscii = require('chartscii');
 
@@ -179,6 +182,9 @@ setInterval(() => createAsciiCharts(), 500);
 
 ## percentage example
 using percentage, solid color with label colors   
+
+<img width="1000" src="https://tool3.github.io/chartscii/img/percentage.svg">   
+
 ```js
 const Chartscii = require('chartscii');
