+++ examples/example.js
@@ -1,0 +1,20 @@
+const Chartscii = require('../index');
+
+// generate random chart data
+const data = [];
+
+for (let i = 1; i <= 20; i++) {
+    data.push(Math.floor(Math.random() * 1000) + 1);
+}
+
+// create chart
+const chart = new Chartscii(data, {
+    label: 'Example Chart',
+    width: 500,
+    sort: true,
+    reverse: true,
+    color: 'pink'
+});
+
+//print chart
+console.log(chart.create());
