+++ examples/fancy.js
@@ -1,4 +1,4 @@
-const Chartscii = require('./index');
+const Chartscii = require('../index');
 
 
 const createAsciiCharts = () => {
@@ -17,8 +17,9 @@ const createAsciiCharts = () => {
     let count = 0;
 
     for (let i = 1; i <= 20; i++) {
+        
         color = colors[Math.floor(Math.random() * colors.length)];
-        data.push({ value: Math.floor(Math.random() * 1000) + 1, label: `${count++}` });
+        data.push({ value: Math.floor(Math.random() * 1000) + 1, color, label: `${count++}` });
     }
 
     // create chart
@@ -28,8 +29,7 @@ const createAsciiCharts = () => {
         sort: true,
         reverse: true,
         color: color,
-        colorLabels: true,
-        percentage: true
+        colorLabels: true
     });
 
     //print chart
@@ -40,4 +40,3 @@ const createAsciiCharts = () => {
 
 
 setInterval(() => createAsciiCharts(), 500);
-
