+++ examples/percentage.js
@@ -1,4 +1,4 @@
-const Chartscii = require('./index');
+const Chartscii = require('../index');
 
 
 const createAsciiCharts = () => {
@@ -17,9 +17,8 @@ const createAsciiCharts = () => {
     let count = 0;
 
     for (let i = 1; i <= 20; i++) {
-        
         color = colors[Math.floor(Math.random() * colors.length)];
-        data.push({ value: Math.floor(Math.random() * 1000) + 1, color, label: `${count++}` });
+        data.push({ value: Math.floor(Math.random() * 1000) + 1, label: `${count++}` });
     }
 
     // create chart
@@ -29,7 +28,8 @@ const createAsciiCharts = () => {
         sort: true,
         reverse: true,
         color: color,
-        colorLabels: true
+        colorLabels: true,
+        percentage: true
     });
 
     //print chart
@@ -40,3 +40,4 @@ const createAsciiCharts = () => {
 
 
 setInterval(() => createAsciiCharts(), 500);
+
