--- percentage.js
@@ -1,43 +1,0 @@
-const Chartscii = require('./index');
-
-
-const createAsciiCharts = () => {
-    let color = '';
-
-    const colors = ['green',
-        'red',
-        'cyan',
-        'pink',
-        'blue',
-        'yellow'
-    ];
-
-    // generate random chart data
-    const data = [];
-    let count = 0;
-
-    for (let i = 1; i <= 20; i++) {
-        color = colors[Math.floor(Math.random() * colors.length)];
-        data.push({ value: Math.floor(Math.random() * 1000) + 1, label: `${count++}` });
-    }
-
-    // create chart
-    const chart = new Chartscii(data, {
-        label: 'Example Chart',
-        width: 500,
-        sort: true,
-        reverse: true,
-        color: color,
-        colorLabels: true,
-        percentage: true
-    });
-
-    //print chart
-    process.stdout.write('\033c');
-    process.stdout.write(`${chart.create()}\n`);
-    
-};
-
-
-setInterval(() => createAsciiCharts(), 500);
-
